pyproject.toml
[tool.poetry]
name = "mini-databae"
version = "0.1.0"
description = ""
authors = ["Ryan <rqn.pham@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uagents = "^0.19.0"
langgraph = "^0.2.69"
langsmith = "^0.3.4"
openai = "^1.61.0"
langchain-groq = "^0.2.4"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.2.0"
langchain-text-splitters = "^0.3.5"
langchain-openai = "^0.3.3"
numpy = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
